python_implementation/requirements.txt
# Django Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
pymongo==4.6.0
mongoengine==0.27.0
redis==5.0.1

# Authentication & Security
django-allauth==0.57.0
django-axes==6.1.1
cryptography==41.0.7
python-jose[cryptography]==3.3.0
webauthn==2.1.0

# File Handling
Pillow==10.1.0
python-magic==0.4.27
django-storages==1.14.2

# Image Processing & Biometrics
opencv-python==4.8.1.78
face-recognition==1.3.0
numpy==1.25.2
scikit-learn==1.3.2

# WebRTC & Real-time
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Background Tasks
celery==5.3.4
celery-beat==2.5.0

# Development & Testing
pytest==7.4.3
pytest-django==4.7.0
coverage==7.3.2
black==23.11.0
flake8==6.1.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0
